051: Extract 2D-Primitives and Random to mixins
