+++ lib/flutter_processing.dart
@@ -2,3 +2,4 @@ library flutter_processing;
 
 export 'src/_core.dart';
 export 'src/color/_color.dart';
+export 'src/shape/shapes.dart';
