+++ lib/src/_core.dart
@@ -1,18 +1,18 @@
 import 'dart:io';
-import 'dart:math';
 import 'dart:typed_data';
 import 'dart:ui';
 
-import 'package:fast_noise/fast_noise.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart' hide Image;
 import 'package:flutter_processing/flutter_processing.dart';
 import 'package:flutter_processing/src/constants/_constants.dart';
+import 'package:flutter_processing/src/math/_random.dart';
 import 'package:flutter_processing/src/math/_trigonometry.dart';
-import 'package:path/path.dart' as path;
+import 'package:flutter_processing/src/shape/_two_d_primitives.dart';
 import 'package:image/image.dart' as imageFormats;
+import 'package:path/path.dart' as path;
 
 import '_painting_context.dart';
 import 'data/conversion.dart';
@@ -313,8 +313,23 @@ class _ProcessingState extends State<Processing> with SingleTickerProviderStateM
   }
 }
 
-class Sketch
-    with SketchConstants, SketchCalculations, SketchTrigonometry, SketchTimeAndDate, SketchConversion, SketchColor {
+abstract class BaseSketch {
+  SketchPaintingContext _paintingContext = SketchPaintingContext();
+
+  @protected
+  SketchPaintingContext get paintingContext => _paintingContext;
+}
+
+class Sketch extends BaseSketch
+    with
+        SketchConstants,
+        SketchMathCalculations,
+        SketchMathTrigonometry,
+        SketchMathRandom,
+        SketchInputTimeAndDate,
+        SketchDataConversion,
+        SketchColor,
+        SketchShapeTwoDPrimitives {
   Sketch.simple({
     Future<void> Function(Sketch)? setup,
     Future<void> Function(Sketch)? draw,
@@ -357,8 +372,6 @@ class Sketch
 
   void Function(Image newFrame)? _onFrameAvailable;
   bool _isDrawing = false;
-
-  SketchPaintingContext _paintingContext = SketchPaintingContext();
 
   bool _hasDoneSetup = false;
 
@@ -562,114 +575,6 @@ class Sketch
     background(color: _backgroundColor);
   }
 
-  //------ Start Random -----
-  Random _random = Random();
-
-  /// Sets the seed value for all [random()] invocations to the given
-  /// [seed].
-  ///
-  /// To return to a natural seed value, pass [null] for [seed].
-  void randomSeed(int? seed) {
-    _random = Random(seed);
-  }
-
-  double random(num bound1, [num? bound2]) {
-    final lowerBound = bound2 != null ? bound1 : 0;
-    final upperBound = bound2 != null ? bound2 : bound1;
-
-    if (upperBound < lowerBound) {
-      throw Exception('random() lower bound must be less than upper bound');
-    }
-
-    return _random.nextDouble() * (upperBound - lowerBound) + lowerBound;
-  }
-
-  int _perlinNoiseSeed = 1337;
-  int _perlinNoiseOctaves = 4;
-  double _perlinNoiseFalloff = 0.5;
-  PerlinNoise? _perlinNoise;
-
-  /// Sets the seed value for all [noise()] invocations to the given
-  /// [seed].
-  ///
-  /// To return to a natural seed value, pass `null` for [seed].
-  void noiseSeed(int? seed) {
-    _perlinNoiseSeed = seed ?? 1337;
-    _initializePerlinNoise();
-  }
-
-  /// Sets the number of [octaves] and the [falloff] for each octave
-  /// for values generated by [noise()].
-  ///
-  /// Omitting a value for [octaves] resets the octaves value to the
-  /// global default.
-  ///
-  /// Omitting a value for [falloff] resets the falloff value to the
-  /// global default.
-  void noiseDetail({
-    int? octaves,
-    double? falloff,
-  }) {
-    _perlinNoiseOctaves = octaves ?? 4;
-    _perlinNoiseFalloff = falloff ?? 0.5;
-
-    _initializePerlinNoise();
-  }
-
-  /// Generates a random value with a Perlin noise algorithm.
-  ///
-  /// Returns values are in [-1.0, 1.0].
-  double noise({
-    required double x,
-    double y = 0,
-    double z = 0,
-  }) {
-    if (_perlinNoise == null) {
-      _initializePerlinNoise();
-    }
-
-    return _perlinNoise!.getPerlin3(x, y, z);
-  }
-
-  void _initializePerlinNoise() {
-    _perlinNoise = PerlinNoise(
-      seed: _perlinNoiseSeed,
-      octaves: _perlinNoiseOctaves,
-      gain: _perlinNoiseFalloff,
-    );
-  }
-
-  double? _previousGaussian;
-
-  /// Returns a `double` from a random series of numbers having the given
-  /// [mean] and the given [standardDeviation].
-  ///
-  /// By default, the [mean] is `0.0`, and the [standardDeviation] is `1`.
-  double randomGaussian({
-    num mean = 0.0,
-    num standardDeviation = 1.0,
-  }) {
-    // The random Gaussian is calculated using the Marsaglia polar method
-    // which generates TWO independent standard normal random variables,
-    // so one is stored and used the next time the function is called.
-    double y1, x1, x2, w;
-    if (_previousGaussian != null) {
-      y1 = _previousGaussian!;
-      _previousGaussian = null;
-    } else {
-      do {
-        x1 = this.random(2) - 1;
-        x2 = this.random(2) - 1;
-        w = x1 * x1 + x2 * x2;
-      } while (w >= 1);
-      w = sqrt(-2 * log(w) / w);
-      y1 = x1 * w;
-      _previousGaussian = x2 * w;
-    }
-
-    return y1 * standardDeviation + mean;
-  }
-
   //------ Start Color/Setting -----
   void background({
     required Color color,
@@ -848,153 +753,6 @@ class Sketch
   }) {
     return ((y * imageWidth) + x) * 4;
   }
-
-  //----- Start Shape/2D Primitives ----
-  void point({
-    required double x,
-    required double y,
-    double? z,
-  }) {
-    if (z != null) {
-      throw UnimplementedError('3D point drawing is not yet supported.');
-    }
-
-    _paintingContext.strokePaint.style = PaintingStyle.fill;
-    _paintingContext.canvas.drawRect(
-      Rect.fromLTWH(x, y, 1, 1),
-      _paintingContext.strokePaint,
-    );
-    _paintingContext.strokePaint.style = PaintingStyle.stroke;
-
-    _paintingContext.markHasUnappliedCanvasCommands();
-  }
-
-  void line(Offset p1, Offset p2, [Offset? p3]) {
-    if (p3 != null) {
-      throw UnimplementedError('3D line drawing is not yet supported.');
-    }
-
-    _paintingContext.canvas.drawLine(p1, p2, _paintingContext.strokePaint);
-
-    _paintingContext.markHasUnappliedCanvasCommands();
-  }
-
-  void circle({
-    required Offset center,
-    required double diameter,
-  }) {
-    _paintingContext.canvas
-      ..drawCircle(center, diameter / 2, _paintingContext.fillPaint)
-      ..drawCircle(center, diameter / 2, _paintingContext.strokePaint);
-
-    _paintingContext.markHasUnappliedCanvasCommands();
-  }
-
-  void ellipse(Ellipse ellipse) {
-    _paintingContext.canvas //
-      ..drawOval(ellipse.rect, _paintingContext.fillPaint) //
-      ..drawOval(ellipse.rect, _paintingContext.strokePaint);
-
-    _paintingContext.markHasUnappliedCanvasCommands();
-  }
-
-  void arc({
-    required Ellipse ellipse,
-    required double startAngle,
-    required double endAngle,
-    ArcMode mode = ArcMode.openStrokePieFill,
-  }) {
-    switch (mode) {
-      case ArcMode.openStrokePieFill:
-        _paintingContext.canvas
-          ..drawArc(ellipse.rect, startAngle, endAngle - startAngle, true, _paintingContext.fillPaint)
-          ..drawArc(ellipse.rect, startAngle, endAngle - startAngle, false, _paintingContext.strokePaint);
-        break;
-      case ArcMode.open:
-        _paintingContext.canvas
-          ..drawArc(ellipse.rect, startAngle, endAngle - startAngle, false, _paintingContext.fillPaint)
-          ..drawArc(ellipse.rect, startAngle, endAngle - startAngle, false, _paintingContext.strokePaint);
-        break;
-      case ArcMode.chord:
-        final chordPath = Path()
-          ..addArc(ellipse.rect, startAngle, endAngle - startAngle)
-          ..close();
-
-        _paintingContext.canvas
-          ..drawArc(ellipse.rect, startAngle, endAngle - startAngle, false, _paintingContext.fillPaint)
-          ..drawPath(chordPath, _paintingContext.strokePaint);
-        break;
-      case ArcMode.pie:
-        _paintingContext.canvas
-          ..drawArc(ellipse.rect, startAngle, endAngle - startAngle, true, _paintingContext.fillPaint)
-          ..drawArc(ellipse.rect, startAngle, endAngle - startAngle, true, _paintingContext.strokePaint);
-        break;
-    }
-
-    _paintingContext.markHasUnappliedCanvasCommands();
-  }
-
-  void square(Square square) {
-    _paintingContext.canvas //
-      ..drawRect(square.rect, _paintingContext.fillPaint) //
-      ..drawRect(square.rect, _paintingContext.strokePaint);
-
-    _paintingContext.markHasUnappliedCanvasCommands();
-  }
-
-  void rect({
-    required Rect rect,
-    BorderRadius? borderRadius,
-  }) {
-    if (borderRadius == null) {
-      _paintingContext.canvas //
-        ..drawRect(rect, _paintingContext.fillPaint) //
-        ..drawRect(rect, _paintingContext.strokePaint);
-    } else {
-      final rrect = RRect.fromRectAndCorners(
-        rect,
-        topLeft: borderRadius.topLeft,
-        topRight: borderRadius.topRight,
-        bottomLeft: borderRadius.bottomLeft,
-        bottomRight: borderRadius.bottomRight,
-      );
-      _paintingContext.canvas //
-        ..drawRRect(rrect, _paintingContext.fillPaint) //
-        ..drawRRect(rrect, _paintingContext.strokePaint);
-    }
-
-    _paintingContext.markHasUnappliedCanvasCommands();
-  }
-
-  void triangle(Offset p1, Offset p2, Offset p3) {
-    final path = Path()
-      ..moveTo(p1.dx, p1.dy)
-      ..lineTo(p2.dx, p2.dy)
-      ..lineTo(p3.dx, p3.dy)
-      ..close();
-
-    _paintingContext.canvas //
-      ..drawPath(path, _paintingContext.fillPaint) //
-      ..drawPath(path, _paintingContext.strokePaint);
-
-    _paintingContext.markHasUnappliedCanvasCommands();
-  }
-
-  void quad(Offset p1, Offset p2, Offset p3, Offset p4) {
-    final path = Path()
-      ..moveTo(p1.dx, p1.dy)
-      ..lineTo(p2.dx, p2.dy)
-      ..lineTo(p3.dx, p3.dy)
-      ..lineTo(p4.dx, p4.dy)
-      ..close();
-
-    _paintingContext.canvas //
-      ..drawPath(path, _paintingContext.fillPaint) //
-      ..drawPath(path, _paintingContext.strokePaint);
-
-    _paintingContext.markHasUnappliedCanvasCommands();
-  }
-  //------- End Shape/2D Primitives -----
 
   //----- Start Shape/Attributes ----
   void strokeWeight(double newWeight) {
@@ -1202,79 +960,6 @@ enum MouseButton {
   left,
   center,
   right,
-}
-
-class Square {
-  Square.fromLTE(Offset topLeft, double extent)
-      : _rect = Rect.fromLTWH(
-          topLeft.dx,
-          topLeft.dy,
-          extent,
-          extent,
-        );
-
-  Square.fromCenter(Offset center, double extent)
-      : _rect = Rect.fromCenter(
-          center: center,
-          width: extent,
-          height: extent,
-        );
-
-  final Rect _rect;
-  Rect get rect => _rect;
-}
-
-class Ellipse {
-  Ellipse.fromLTWH({
-    required Offset topLeft,
-    required double width,
-    required double height,
-  }) : _rect = Rect.fromLTWH(
-          topLeft.dx,
-          topLeft.dy,
-          width,
-          height,
-        );
-
-  Ellipse.fromLTRB({
-    required Offset topLeft,
-    required Offset bottomRight,
-  }) : _rect = Rect.fromLTRB(
-          topLeft.dx,
-          topLeft.dy,
-          bottomRight.dx,
-          bottomRight.dy,
-        );
-
-  Ellipse.fromCenter({
-    required Offset center,
-    required double width,
-    required double height,
-  }) : _rect = Rect.fromCenter(
-          center: center,
-          width: width,
-          height: height,
-        );
-
-  Ellipse.fromCenterWithRadius({
-    required Offset center,
-    required double radius1,
-    required double radius2,
-  }) : _rect = Rect.fromCenter(
-          center: center,
-          width: radius1 * 2,
-          height: radius2 * 2,
-        );
-
-  final Rect _rect;
-  Rect get rect => _rect;
-}
-
-enum ArcMode {
-  openStrokePieFill,
-  open,
-  chord,
-  pie,
 }
 
 enum ImageFileFormat {
