+++ lib/src/shape/shapes.dart
@@ -1,0 +1,74 @@
+import 'dart:ui';
+
+class Square {
+  Square.fromLTE(Offset topLeft, double extent)
+      : _rect = Rect.fromLTWH(
+          topLeft.dx,
+          topLeft.dy,
+          extent,
+          extent,
+        );
+
+  Square.fromCenter(Offset center, double extent)
+      : _rect = Rect.fromCenter(
+          center: center,
+          width: extent,
+          height: extent,
+        );
+
+  final Rect _rect;
+  Rect get rect => _rect;
+}
+
+class Ellipse {
+  Ellipse.fromLTWH({
+    required Offset topLeft,
+    required double width,
+    required double height,
+  }) : _rect = Rect.fromLTWH(
+          topLeft.dx,
+          topLeft.dy,
+          width,
+          height,
+        );
+
+  Ellipse.fromLTRB({
+    required Offset topLeft,
+    required Offset bottomRight,
+  }) : _rect = Rect.fromLTRB(
+          topLeft.dx,
+          topLeft.dy,
+          bottomRight.dx,
+          bottomRight.dy,
+        );
+
+  Ellipse.fromCenter({
+    required Offset center,
+    required double width,
+    required double height,
+  }) : _rect = Rect.fromCenter(
+          center: center,
+          width: width,
+          height: height,
+        );
+
+  Ellipse.fromCenterWithRadius({
+    required Offset center,
+    required double radius1,
+    required double radius2,
+  }) : _rect = Rect.fromCenter(
+          center: center,
+          width: radius1 * 2,
+          height: radius2 * 2,
+        );
+
+  final Rect _rect;
+  Rect get rect => _rect;
+}
+
+enum ArcMode {
+  openStrokePieFill,
+  open,
+  chord,
+  pie,
+}
